use the full resolution we get from the WMA decoder and let the DSP code do the rest
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@14224 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/libwma/wmadec.h
@@ -172,6 +172,6 @@ int wma_decode_init(WMADecodeContext* s, asf_waveformatex_t *wfx);
 int wma_decode_superframe_init(WMADecodeContext* s,
                                uint8_t *buf, int buf_size);
 int wma_decode_superframe_frame(WMADecodeContext* s,
-                                int16_t *samples,
+                                int32_t *samples,
                                 uint8_t *buf, int buf_size);
 #endif
